--- Flask backend/__init__.py
@@ -1,10 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-from flask_jwt_extended import JWTManager
-from flask_login import LoginManager
-from flask_cors import CORS
-
-db = SQLAlchemy()
-bcrypt = Bcrypt()
-jwt = JWTManager()
-login_manager = LoginManager()
