--- Flask backend/app.py
@@ -1,38 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-from flask_jwt_extended import JWTManager
-from flask_login import LoginManager
-from flask_cors import CORS
-import os
-
-# Import application configuration and routes
-from config import Config
-from routes import register_routes
-
-# Initialize extensions
-db = SQLAlchemy()
-bcrypt = Bcrypt()
-jwt = JWTManager()
-login_manager = LoginManager()
-
-def create_app():
-    # Initialize Flask app
-    app = Flask(__name__)
-
-    # Configure app
-    app.config.from_object(Config)
-
-    # Initialize extensions with app
-    db.init_app(app)
-    bcrypt.init_app(app)
-    jwt.init_app(app)
-    login_manager.init_app(app)
-
-    # Enable CORS
-    CORS(app, resources={r"/*": {"origins": "http://localhost:5173"}}, supports_credentials=True)
-
-    # Register routes
-    register_routes(app)
-
-    return app
