--- Flask backend/full_database_dump.sql
@@ -1,3 +1,0 @@
-PRAGMA foreign_keys=OFF;
-BEGIN TRANSACTION;
-COMMIT;
