+++ Flask backend/routes/auth_routes.py
@@ -1,0 +1,49 @@
+from flask import Blueprint, request, jsonify
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt
+from config import bcrypt, blacklist
+from models import get_users_collection
+
+auth_bp = Blueprint('auth', __name__)
+
+@auth_bp.route('/login', methods=['POST'])
+def login():
+    data = request.get_json()
+    email = data.get('email')
+    password = data.get('password')
+
+    users = get_users_collection()
+    user = users.find_one({'email': email})
+
+    if user and bcrypt.check_password_hash(user['password'], password):
+        access_token = create_access_token(identity=str(user['_id']))
+        return jsonify({'message': 'Login successful', 'token': access_token}), 200
+    else:
+        return jsonify({'message': 'Invalid credentials'}), 401
+
+@auth_bp.route('/register', methods=['POST'])
+def register():
+    data = request.get_json()
+    email = data.get('email')
+    username = data.get('username')
+    password = data.get('password')
+
+    users = get_users_collection()
+    if users.find_one({'email': email}):
+        return jsonify({'message': 'User already exists'}), 400
+
+    hashed_password = bcrypt.generate_password_hash(password).decode('utf-8')
+    user = {
+        'email': email,
+        'username': username,
+        'password': hashed_password
+    }
+    users.insert_one(user)
+
+    return jsonify({'message': 'User registered successfully'}), 201
+
+@auth_bp.route('/logout', methods=['POST'])
+@jwt_required()
+def logout():
+    jti = get_jwt()["jti"]
+    blacklist.add(jti)
+    return jsonify({"message": "Successfully logged out"}), 200
